Move HandCardsBlock
Move HandCardsBlock into new front end frame.
